requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
cryptography==46.0.4
fastapi==0.128.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.2
packaging==26.0
pandas==3.0.0
pandera==0.29.0
pluggy==1.6.0
propcache==0.4.1
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
questdb==4.1.0
six==1.17.0
starlette==0.50.0
typeguard==4.4.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
voluptuous==0.16.0
websockets==16.0
yarl==1.22.0
